Maximal network planner
This adds a ligand network planner to create as many edges as the
mappers can propose. This isn't the same as all-to-all, since some
changes might be beyond that a mapper can create a mapping for, but it
will be as close to all-to-all as we can propose.

It will be particularly useful in developing networks and in debugging.
Already, this is really just a refactor of stuff that had been part of
the minimal spanning graph.