context manager to silence root logging
This is designed to wrap imports that use `logging.basiConfig`.

`logging.basicConfig` is aimed at users. Libraries shouldn't use
it. This temporarily attaches a handler to the root logger (which
makes `basicConfig` a no-op). It then resets the root logger
handlers, in case someone in user-land had set them previously.