Explanation docs (initial work)
As discussed in the meeting today, this is a rough outline of the
"explanation" docs (per diataxis). The approach I've been taking is to
think of it as stripping down layers of abstraction, so we start out
with the very abstract stuff and get more into the details as we go on.

Basically, the goal is "here's the simplest way to do it -- if that's
too simple for your needs, here's the next simplest." These docs should
also have lots of links to how-tos as we add them; the goal here is to
start from the big picture.

This also includes docs for CLI usage.