docs: update badge on front page
This badge should point to stable/released docs, so people are linked to what documents the thing that they are most likely to be using.  Most people will be using conda builds, not dev/source builds, if they're using dev builds they'll also know how to get to the dev docs.

Removed disclaimer on pre-alpha status, we're 1.0rc