Fix test to check if we are printing `--allow-partial` when a user uses `--allow-partial` (#1564)
* escape string correctly

* Fix test to check if string is printed

This test used to use pytest.warns to check if we printed a string to
the terminal. This isn't the right way to test if things get printed to
terminal since it only checks warnings raised in python-land. The test
passed since the regex used to match warnings ended up matching warnings
raised since it was a negated character class which matches a single
character that is not any of the characters listed inside.